Added project awareness
Luacheck will now be able to find and lint the project directory
containing a .luacheckrc file. If the option is disabled, ALE is able to
return to linting the current buffer as before.

Additionally, ALE will try to navigate to the project directory before
linting if a .luacheckrc is found.