Fix golangci-lint to run on package and not files.
I don't think golangci-lint ever operated on individual files.

Like most go tooling it operates on packages. A package is just a
collection of files but also has an "import path" which is used
to identify on the filesystem where things are (using GOPATH or
the go modules system).

Before this change ALE would run the following on a buffer viewing
foo.go.

    cd /path/to/foo/ && golangci-lint run foo.go --options

This change omits the file in the buffer and just runs it on the
directory; golangci-lint will automatically figure out the package
import.

    cd /path/to/foo && golangci-lint run --options

I also removed traces of g:ale_go_golangci_lint_package since
this shouldn't be a choice.

If and when golangci-lint decides to implement some way to
operate on indviidual files, then we should consider introducing
it again and making package scope the default. Wouldn't
worry about the performance implications of this, golangci-lint
is pretty damn fast at package scope.