Add a language option "-l" to known filetypes for uncrustify
This change adds a `-l <language>` parameter to the final uncrustify
command that is executed via ale_fix().

ALE passes the current buffer to uncrustify via an input stream without
providing a language parameter. This causes an issue where some of the
uncrustify configuration options are ignored and the output is different
than expected.

This change introduces a global `g:ale_c_uncrustify_language_mappings`
that is a dictionary of mappings between vim filetypes and uncrustify
recognized languages.