Run `make -Bn` instead of `make -n`
Currently, `make -n` is run to obtain the flags. 
But if the default target is already built, make will exit saying something like `make: Nothing to be done for 'all'.`
By using `-B` flag which is short for `--always-build`, make will try to build every file and prepare command lines for them.
Since we also specify `-n` which is short for `--dry-run`, no commands will be executed.
Instead all the commands that were to be executed will be printed, which is then used by this script to determine the `CFLAGS`.