WIP: LSP: Support workspace/configuration
According to
https://github.com/microsoft/language-server-protocol/issues/676 the
didChangeConfiguration method only triggers a workspace/configuration
from the server, which the client needs to handle, so effectively the
current configuration support in ALE is not working at all with modern
LSP servers.

This is a work-in-progress to add support for workspace/configuration,
currently only hacked together far enough that it works for my use with
rust-analyzer. I hope to extend it / clean it up, but if anyone wants to
pick up from here, feel free to.