Add option to sync buffer cursor with loc|qf list.
Pressing enter in the loclist or quickfix list jumps to the
corresponding line in the buffer.

This MR introduces a new option `g:ale_follow_list` that allows
ALE to do the opposite. Moving the cursor inside the buffer, if the
current line has a linter error, then it will move the cursor inside
the loc|qf list to the line where that error is.