code_action: Fix EOL at EOF corner cases while performing no changes
Instead of always removing the trailing empty line and hoping it wasn't
supposed to be there, only remove it when we know ale#util#Writefile
will add a trailing newline, and additionally tell ale#util#Writefile to
not add that newline if any change explicitly deleted that newline.

(Unfortunately with a:should_save=0, adding a trailing newline that
wasn't there results in noeol and empty line at end of buffer, because
for some reason setbufvar(l:buffer, '&eol', 1) doesn't work in buffers
that started as noeol. This is likely a bug in vim itself.)

Includes a minor fix that prevents invoking ale#util#SetBufferContents
without a buffer. Not sure whether that might really happen anywhere.

(test_code_action_python.vader depends on two incorrect behaviours
cancelling each other, so it's temporarily adjusted to pass)