code_action: Fix column around EOL corner cases
The original code didn't allow changing the last character in a line
without also replacing the newline, due to an off-by-one error.

Or, from a different point of view, it didn't allow replacing the
newline without also replacing the preceding character. Therefore it
comes as no surprise that it didn't correctly handle whether to replace
or leave the newline at the end of the changed range.

This commit fixes both, and also adjusts two tests that depended on the
incorrect behaviour. Also, the temporary change to
test_code_action_python.vader introduced earlier can now be reverted.

Finally, this adds a _lot_ of tiny unit tests for various corner cases.
These were verified against the reference implementation in
[vscode-languageserver-node][], and a javascript source to run them is
added in a followup commit. The added tests take an additional second or
two. It might be worth reducing them somewhat to only test what was
broken…?

[vscode-languageserver-node]: https://github.com/microsoft/vscode-languageserver-node/