Fix breakage w/ plugins that inadvertently trigger ALE in `execute()`
Vim does not allow the use of `:redir` from within `execute()` calls, no
matter how deeply within the call chain. There are, sadly, plugins which
call `execute()` for whatever reason and inadvertently trigger functions
that use `:redir`, thereby raising an `E930`:

>E930: Cannot use :redir inside execute()

We replace all uses of `:redir => var` with `let var = execute(…)`, that
resolves this.

---

In `ale#sign#SetUpDefaultColumnWithoutErrorsHighlight()`, the use of the
`0verbose` prefix had to be removed, as `execute()` doesn't accept it in
its second argument. In fact:

>execute({command} [, {silent}])					*execute()*
>
>[…]
>
>		The optional {silent} argument can have these values:
>			""		no `:silent` used
>			"silent"	`:silent` used
>			"silent!"	`:silent!` used

Given that `verbose` is already set to `0` on the previous line, the use
of `0verbose` seems unnecessary.

>function! ale#sign#SetUpDefaultColumnWithoutErrorsHighlight() abort
>    let l:verbose = &verbose
>    set verbose=0
>    redir => l:output
>        0verbose silent highlight SignColumn
>    redir end
>    let l:output = execute('highlight SignColumn', 'silent')
>    let &verbose = l:verbose
>[…]

The history of this part of the code shows that the `0verbose` was added
first, then the let-verbose-set-verbose-let-verbose dance. I believe the
`0verbose`, after that part was added, became unnecessary, but it wasn't
removed. The commit which added the dance is bafe1c0, and the `0verbose`
was added in d3ed1e5.

Vim docs on `:[count]verbose` say:

>:[count]verb[ose] {command}
>			Execute {command} with 'verbose' set to [count].  If
>			[count] is omitted one is used. ":0verbose" can be
>			used to set 'verbose' to zero.