ALERename: Fix an issue with unlisted buffers
After calling ALERename on certain symbols, some files would be
incorrectly replaced. The whole file would be truncated, and the only
contents would be the new variable name.

I was able to track it down to the following configuration:

    let g:ale_linters = {
    \  'go': ['gopls', 'golangci-lint'],
    \}

Whenever I tried to rename a symbol in `myfile.go`, which was referenced from an unopened file `myfile_test.go`, the whole `myfile_test.go` would be truncated.

I tracked it down to `bufnr('/path/to/myfile_test.go')` returning number
for a buffer that I never opened.

I later noticed that this buffer was unlisted (visible only via :ls!).
So there must be something in the golangci-lint linter that opens this
file in the background.

This fix adds a check for hidden buffers before applying code actions.

Closes #3781