Fix cases where echon fails
It seems that if the string is too wide for the window, `echon` fails
silently?  Or just displays nothing.  This truncates the string like in
the error catch, although it does now make you wonder what's the point
of the try/catch in the first place.

I reverted some of my changes for the catch statement, to minimize
changes, and at least give some difference.  I am not sure about any of
the original thinking for this, and I don't know a lot about writing
VimL.

One thing I did notice - I neede `-5` instead of `-4`, not sure if we
should fix the catch as well.