Re-apply #3538 - remove -T argument from ruby linter (#3951)
The `-T` option (for "taint checking") was deprecated in ruby 2.7
and removed entirely in ruby 3.0. This causes the linter to fail
entirely for users of ruby 3.0.

This was reported in #3537, and then fixed in #3538 - but it seems as
though in 9fe7b1fe6a23fb55e6d782293585d58193123f59, it was accidentally
and entirely undone.

This commit is essentially identical to #3538, aside from a path change
for the tests.