Make run-tests compatible with Podman
There is no need to filter for references in such a complicated way.
docker images already works if you just pass the image and tag as
an argument.

This caused problems if one was using podman with its docker-compatible
interface.
Previously podman would return the following error:
Error: cannot specify an image and a filter(s)

With this new method podman does not return an error anymore, causing
the image to not be redownloaded every time and it still works with
normal Docker.