Don't use temp file for clj-kondo
clj-kondo since version 2022.03.04 has a check to for the namespace
declared in the file matching the actual file hierarchy, which now
always fails due to the file hierarchy generated for %t not matching the
file to be linted.  Since clj-kondo only reads the file, use %s instead
of %t to start it.

See: https://github.com/clj-kondo/clj-kondo/issues/1240