Implement ALEHover/ale_hover_cursor callbacks
Proof of concept for user-defined display functions.  User can pass a
FuncRef to ale#hover#SetCallback to define behavior for explicit calls
to ALEHover, or to ale#hover#SetCursorCallback for cursor idle when
ale_hover_cursor is set.

The callback will be passed a dictionary of information, which contain
the fields "lines" and "commands".