Fix virtual text for vim 8.2
Regression was introduced in d93bc2baf7532818e83bf2fac61fcd591beb6151

The problem was that we did not handle the edge case where there is no
last popup to close, which caused old vim versions to enter code by
accident that was only supposed to be run by vim 9.

We fix this by guarding the if condition for vim 9.

Fixes #4290