Correctly test g:ale_virtualtext_cursor
The change made for #4475 did not work correctly because when Vim
compares a non-numeric string with a number using !=, it converts the
string to 0 before comparing.  Thus the second half of the clause
(g:ale_virtualtext_cursor != 0) was always false if
g:ale_virtualtext_cursor was 'all' or 'current'.  Fix by using isnot#,
which does not coerce strings to numbers when comparing.