Fix project file search for haskell hls
Previously, we would append the candidate root files to the output paths
from `ale#path#Upwards(expand(#buffer:p:h))` which do not include a
trailing forwardslash (/).
These would form invalid paths which would never trigger `filereadable`.

Furthermore, we need to return `l:path . '/'` because in
`ale#haskell#hls#GetProjectRoot` we take `:h` based on the project root
file. In the case a root file was found, we want to keep the directory
it was found in, so we need to identify the path as a directory (again,
with the forward slash)

The correct identification of root files allows us to use LSP in multi
package projects, and in projects with a custom hie.yaml, such as GHC!