go/golangci_lint.vim: set package linting to 1
The present default of 0 causes excessive and unnecessary typecheck errors in all but the most trivial (single-file) packages. The standard Go unit of code is a directory, not a file, and optimizing for processing a single file provides negligible savings in processing time for a typical project, yet creates a lot of noise for the user.