fix(eslint): handle monorepo setup for eslint
The heuristic for finding the correct Cwd when executing eslint does not
work in the context of a monorepo setup where the `.eslintrc.js` may not
live in the root of the repo but in one of the leaf packages ie. each
package may have its own settings.

This PR fixes the behaviour by setting Cwd to the closest config path
instead of the bin path.